requirements.txt
python-dotenv
openai
flask
requests
pandas
scikit-learn
nltk
gunicorn